pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nexus-atom"
version = "0.1.0"
description = "Spec-driven engine with deterministic atom lifecycle management"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "networkx>=3.0",
    "typer>=0.9",
    "python-frontmatter>=1.0",
    "pyyaml>=6.0",
]

[project.scripts]
nexus-atom = "cli.main:app"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "mypy>=1.0",
    "ruff>=0.1",
    "types-PyYAML",
]

[tool.hatch.build.targets.wheel]
packages = ["src/nexus_atom", "cli"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src", "."]

[tool.mypy]
strict = true
python_version = "3.11"
files = ["src", "cli", "tests"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "ANN"]
